src/requirements.txt
# Dependencies for the dashboard (src/app.py). Used by Streamlit Community Cloud when entrypoint is src/app.py.
streamlit>=1.28.0
supabase>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
streamlit-searchbox>=0.1.0
